How to use defaultdict
A beginner program to help understand using defaultdict from collections through import csv.  This exercise demonstrates the easy creation of a dictionary build lists of cities as values for state codes as keys.
Also included is the use of additional data handling terms for iterating, checking values, sorting and printing.